@@ -1,9 +1,9 @@
-class UserRoles < ActiveRecord::Migration
+class UserRoles < ActiveRecord::Migration[5.0]
   def up
     create_table :roles do |t|
       t.string :name
     end
-    create_table :roles_users, :id => false do |t|
+    create_table :roles_users, id: false do |t|
       t.references :role
       t.references :user
     end
